fixed connection issues and renamed utils file
