+++ JavaAPI/src/main/java/services/AssessmentService.java
@@ -6,12 +6,8 @@ import models.Assessment;
 import models.Grade;
 import models.Note;
 import models.AssessmentType;
-import util.dbconnection;
 
-import java.sql.PreparedStatement;
-import java.sql.ResultSet;
 import java.sql.SQLException;
-import java.util.ArrayList;
 import java.util.List;
 
 public class AssessmentService {
