+++ JavaAPI/src/test/java/testAssignAssessmentType.java
@@ -1,22 +1,16 @@
-import dao.AssessmentDAO;
 import dao.AssessmentDAOImpl;
-import models.Assessment;
 import models.Grade;
 import org.junit.Before;
 import org.junit.Test;
 import org.junit.jupiter.api.AfterAll;
 import org.junit.jupiter.api.AfterEach;
-import org.junit.jupiter.api.BeforeAll;
-import org.junit.jupiter.api.BeforeEach;
 import org.mockito.MockedStatic;
 import org.mockito.Mockito;
-import util.dbconnection;
+import util_project.dbconnection;
 
 import java.sql.Connection;
 import java.sql.PreparedStatement;
 import java.sql.ResultSet;
-import java.sql.SQLException;
-import java.util.ArrayList;
 
 import static org.junit.jupiter.api.Assertions.*;
 
@@ -54,8 +48,8 @@ public class testAssignAssessmentType {
 
     @Test
     public void testNotNull() throws Exception {
-       // Boolean returnedAssessmentType = adao.assignAssessmentType(1);
-        //assertNotNull(returnedAssessmentType);
+        Boolean returnedAssessmentType = adao.assignAssessmentType(1, 1);
+        assertNotNull(returnedAssessmentType);
     }
 
     @Test
