+++ JavaAPI/src/test/java/testGetAssessments.java
@@ -1,25 +1,19 @@
-import dao.AssessmentDAO;
 import dao.AssessmentDAOImpl;
 import models.Assessment;
 import models.Grade;
 import org.junit.Before;
 import org.junit.Test;
 import java.util.List;
-import org.junit.jupiter.api.*;
+
 import org.junit.jupiter.api.AfterAll;
 import org.junit.jupiter.api.AfterEach;
-import org.junit.jupiter.api.BeforeAll;
-import org.junit.jupiter.api.BeforeEach;
-import org.mockito.Mock;
 import org.mockito.MockedStatic;
 import org.mockito.Mockito;
-import org.mockito.MockitoAnnotations;
-import util.dbconnection;
+import util_project.dbconnection;
 
 import java.sql.Connection;
 import java.sql.PreparedStatement;
 import java.sql.ResultSet;
-import java.sql.SQLException;
 import java.util.ArrayList;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
